Update to use standard fine tuning

+++ latent_editing/latent_pipeline.py
@@ -88,43 +88,23 @@ def run_attribution_steps(
 
 
 def identify_target_components(attribution_scores: dict):
-    ig_rewrite_original_scores = attribution_scores[
-        AttributionMethod.IG_REWRITE_ORIGINAL
-    ]
-    ig_original_rewrite_scores = attribution_scores[
-        AttributionMethod.IG_ORIGINAL_REWRITE
-    ]
-    # ap_highlighted = highlighted_dict[AttributionMethod.AP_ORIGINAL_REWRITE]
-
-    # Identify minimal components as those with high attribution scores in both IG and AP.
-    # minimal_components = ig_rewrite_original_highlighted & ap_highlighted
+    ig_rewrite_original = attribution_scores[AttributionMethod.IG_REWRITE_ORIGINAL]
+    ig_original_rewrite = attribution_scores[AttributionMethod.IG_ORIGINAL_REWRITE]
+    # ap_highlighted = attribution_scores[AttributionMethod.AP_ORIGINAL_REWRITE]
 
     # Identify latent components as those with high attribution scores in only one direction of IG.
-    asymmetry_scores = asymmetry_score(ig_rewrite_original_scores, ig_original_rewrite_scores, is_ig=True)
+    asymmetry_scores = asymmetry_score(ig_rewrite_original, ig_original_rewrite, is_ig=True)
     latent_components = highlight_components(asymmetry_scores)[0]
-    return latent_components
 
-    # Get union of minimal and latent components
-    # return minimal_components | latent_components
+    important_components = highlight_components(ig_rewrite_original)[0]
+    target_components = important_components | latent_components
 
-
-def inverted_hinge_loss(output_logits, target_index):
-    logit_probs = torch.softmax(output_logits, dim=-1)
-    # Get probability of target token for each sample
-    # target_prob = torch.gather(logit_probs, dim=1, index=target_index.unsqueeze(1))
-    target_prob = logit_probs[:, target_index]
-    # Get max probability of non-target tokens
-    nontarget_probs = logit_probs.clone()
-    nontarget_probs[:, target_index] = -math.inf
-    max_nontarget_prob = torch.max(nontarget_probs, dim=-1)[0]
-    # Calculate IHL
-    return (1 + target_prob - max_nontarget_prob).sum()
+    return target_components
 
 
 def optimise_edit_components(
     model: HookedTransformer,
-    forget_logits: Tensor,
-    retain_logits: Tensor,
+    logits: Tensor,
     answer_indices: Tensor,
     target_mlp_components: Tensor,
     target_attn_components: Tensor,
@@ -136,19 +116,11 @@ def optimise_edit_components(
     """
     optimiser.zero_grad()
 
-    # Decrease prediction score on original token
-    loss_forget = inverted_hinge_loss(forget_logits, answer_indices[:, 0])
-    # Make prediction probability distribution similar to predictions given rewritten prompt
-    loss_rewrite = F.cross_entropy(retain_logits, answer_indices[:, 1], reduction="sum")
-
-    kl_loss = torch.nn.KLDivLoss(reduction="batchmean")
-    forget_log_prob = F.log_softmax(forget_logits, dim=1)
-    rewrite_prob = F.softmax(retain_logits, dim=1)
-    loss_fluency = kl_loss(forget_log_prob, rewrite_prob)
-
-    # Because the rewrite output might not match the rewrite target token, we do not try to match the probability distribution exactly
-    loss = loss_forget + loss_rewrite + 0.5 * loss_fluency
-    print(f"Total loss: {loss}, forget loss: {loss_forget}, rewrite loss: {loss_rewrite}, fluency loss: {loss_fluency}")
+    # Fine tune on conditional likelihood of edit target given the original prompt
+    edit_target = answer_indices[:, 1].unsqueeze(1)
+    log_probs = -torch.nn.functional.log_softmax(logits, dim=-1)
+    nll_loss = log_probs.gather(dim=-1, index=edit_target)
+    loss = nll_loss.mean()
 
     loss.backward()
 
@@ -204,6 +176,7 @@ def localise_model(
     rewrite_prompts: list[str],
     answer_labels: Tensor,
     sample_index: int,
+    overwrite: bool = False,
 ):
     assert len(original_prompts) == len(rewrite_prompts), f"Must have same number of prompts"
     n_samples = len(original_prompts)
@@ -235,24 +208,28 @@ def localise_model(
 
     # if not is_mlp_saved and not is_mlp_saved:
 
-    mlp_attribution_scores, attn_attribution_scores = run_attribution_steps(
-        model,
-        original_tokens,
-        rewrite_tokens,
-        answer_labels,
-        original_cache,
-        rewrite_cache,
-        original_logit_diff,
-        rewrite_logit_diff
-    )
+    if overwrite:
+        mlp_attribution_scores, attn_attribution_scores = run_attribution_steps(
+            model,
+            original_tokens,
+            rewrite_tokens,
+            answer_labels,
+            original_cache,
+            rewrite_cache,
+            original_logit_diff,
+            rewrite_logit_diff
+        )
 
-    target_mlp = identify_target_components(mlp_attribution_scores).to(model.cfg.device)
-    target_attn = identify_target_components(attn_attribution_scores).to(model.cfg.device)
+        target_mlp = identify_target_components(mlp_attribution_scores).to(model.cfg.device)
+        target_attn = identify_target_components(attn_attribution_scores).to(model.cfg.device)
 
-    torch.save(target_mlp, target_mlp_save_path)
-    torch.save(target_attn, target_attn_save_path)
+        torch.save(target_mlp, target_mlp_save_path)
+        torch.save(target_attn, target_attn_save_path)
+    else:
+        target_mlp = torch.load(target_mlp_save_path)
+        target_attn = torch.load(target_attn_save_path)
 
-    return target_mlp[0], target_attn[0]
+    return target_mlp, target_attn
 
 
 def edit_model(
@@ -260,6 +237,7 @@ def edit_model(
     original_prompt: str,
     rewrite_prompt: str,
     answer_labels: Tensor,
+    paraphrased: list[str],
     target_mlp: Tensor,
     target_attn: Tensor,
     localise: bool = True,
@@ -269,20 +247,28 @@ def edit_model(
     relevant_parameters = [
         p for name, p in model_copy.named_parameters() if "attn" in name or "mlp" in name
     ]
-    optimiser = optim.AdamW(relevant_parameters, lr=3e-4)
+    optimiser = optim.AdamW(relevant_parameters, lr=6e-5)
     
     # Fine tune until loss is below threshold
-    loss = math.inf
-    max_epochs = 10
-    n = 0
-    while loss > 1.5 and n < max_epochs:
-        forget_logits = model_copy(original_prompt)[:, -1, :]
-        rewrite_logits = model_copy(rewrite_prompt)[:, -1, :]
-        # answer_index = answer_labels[i, 1].unsqueeze(0)  # Aim for rewritten answer
-        
+    max_epochs = 5
+    for i in range(max_epochs):
+        logits = model_copy(original_prompt)[:, -1, :]
         loss = optimise_edit_components(
-            model_copy, forget_logits, rewrite_logits, answer_labels, target_mlp, target_attn, optimiser, localise
+            model_copy, logits, answer_labels, target_mlp, target_attn, optimiser, localise
         )
-        n += 1
+
+        paraphrased_logits = model_copy(paraphrased)[:, -1, :]
+        paraphrased_loss = optimise_edit_components(
+            model_copy, paraphrased_logits, answer_labels, target_mlp, target_attn, optimiser, localise
+        )
+
+        rewrite_logits = model_copy(rewrite_prompt)[:, -1, :]
+        rewrite_loss = optimise_edit_components(
+            model_copy, rewrite_logits, answer_labels, target_mlp, target_attn, optimiser, localise
+        )
+        
+        print(
+            f"Epoch {i + 1}/{max_epochs} - Loss: {loss.item():.4f}, Paraphrased Loss: {paraphrased_loss.item():.4f}, Random Loss: {rewrite_loss.item():.4f}"
+        )
 
     return model_copy
